Combine and move the felt crates to browser/extensions/felt
