+++ browser/extensions/felt/rust/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "felt"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+log = "0.4"
+env_logger = { version = "0.10", default-features = false }
+xpcom = { path = "../../../../xpcom/rust/xpcom" }
+nserror = { path = "../../../../xpcom/rust/nserror" }
+cookie = "0.16"
+cstr = "0.2"
+ipc-channel = "0.19"
+moz_task = { path = "../../../../xpcom/rust/moz_task" }
+nsstring = { path = "../../../../xpcom/rust/nsstring" }
+serde = { version = "1.0.219", features = ["derive"] }
+thin-vec = "0.2.12"
+time = "0.3.36"
+serde_json = "1.0"
