+++ browser/extensions/felt/rust/components.conf
@@ -12,7 +12,7 @@ Classes = [
     {
         'cid': '{53c53451-3b93-440c-b27b-9759ccb29da6}',
         'contract_ids': ['@mozilla.org/toolkit/library/felt;1'],
-        'headers': ['/toolkit/library/felt/src/felt.h'],
+        'headers': ['mozilla/browser/extensions/felt/felt.h'],
         'legacy_constructor': 'felt_constructor',
         'js_name': 'felt',
         'singleton': True,
@@ -21,7 +21,7 @@ Classes = [
     {
         'cid': '{03820028-493d-4cf0-812d-828fb68d3f1e}',
         'contract_ids': ['@mozilla-org/felt-restart-forced;1'],
-        'headers': ['/toolkit/library/felt/src/felt.h'],
+        'headers': ['mozilla/browser/extensions/felt/felt.h'],
         'categories': {'content-policy': '@mozilla-org/felt-restart-forced;1'},
         'legacy_constructor': 'felt_restartforced_constructor',
         'interfaces': ['nsIFeltRestartForced'],
