+++ browser/extensions/felt/rust/moz.build
@@ -4,8 +4,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-EXPORTS.mozilla.toolkit.library += [
-    "felt_ffi.h",
+EXPORTS.mozilla.browser.extensions.felt += [
+    "felt.h",
 ]
 
 XPIDL_SOURCES += [
