+++ browser/extensions/felt/rust/nsIFeltRestartForced.idl
