+++ browser/extensions/felt/rust/src/lib.rs
@@ -3,15 +3,27 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 use log::trace;
-use std::ffi::CStr;
 use std::os::raw::c_char;
+use std::{ffi::CStr, sync::atomic::AtomicBool};
 
 use std::env;
 use std::sync::{atomic::Ordering, Mutex};
 
-use felt;
+#[macro_use]
+extern crate cstr;
+#[macro_use]
+extern crate xpcom;
+extern crate thin_vec;
+
+mod client;
+mod components;
+mod message;
+mod utils;
 
 use env_logger;
+
+static IS_FELT_UI: AtomicBool = AtomicBool::new(false);
+static IS_FELT_BROWSER: AtomicBool = AtomicBool::new(false);
 
 #[no_mangle]
 pub extern "C" fn felt_init() {
@@ -29,18 +41,18 @@ pub extern "C" fn felt_init() {
 
     let is_felt_ui = found_felt_ui_arg || found_felt_ui_env;
     trace!("felt_init(): is_felt_ui={}", is_felt_ui);
-    felt::IS_FELT_UI.store(is_felt_ui, Ordering::Relaxed);
+    IS_FELT_UI.store(is_felt_ui, Ordering::Relaxed);
 
     let is_felt_browser = env::args()
         .into_iter()
         .any(|arg| arg.replace("-", "").replace("/", "").to_lowercase() == "felt");
 
     trace!("felt_init(): is_felt_browser={}", is_felt_browser);
-    felt::IS_FELT_BROWSER.store(is_felt_browser, Ordering::Relaxed);
+    IS_FELT_BROWSER.store(is_felt_browser, Ordering::Relaxed);
 
     assert!(
         !(is_felt_browser && is_felt_ui),
-        "Cannot have both -fletUI and -felt args"
+        "Cannot have both -feltUI and -felt args"
     );
 
     trace!("felt_init() done");
@@ -49,23 +61,23 @@ pub extern "C" fn felt_init() {
 #[no_mangle]
 pub extern "C" fn is_felt_ui() -> bool {
     trace!("is_felt_ui()");
-    felt::IS_FELT_UI.load(Ordering::Relaxed)
+    IS_FELT_UI.load(Ordering::Relaxed)
 }
 
 #[no_mangle]
 pub extern "C" fn is_felt_browser() -> bool {
     trace!("is_felt_browser()");
-    felt::IS_FELT_BROWSER.load(Ordering::Relaxed)
+    IS_FELT_BROWSER.load(Ordering::Relaxed)
 }
 
-pub static FELT_CLIENT: Mutex<Option<felt::FeltClientThread>> = Mutex::new(None);
+pub static FELT_CLIENT: Mutex<Option<client::FeltClientThread>> = Mutex::new(None);
 
 #[no_mangle]
 pub extern "C" fn firefox_connect_to_felt(server_name: *const c_char) -> () {
     let srv_name = unsafe { CStr::from_ptr(server_name) };
     let server_socket = String::from_utf8_lossy(srv_name.to_bytes()).to_string();
     trace!("firefox_connect_to_felt({})", server_socket);
-    match felt::FeltClientThread::new(server_socket) {
+    match client::FeltClientThread::new(server_socket) {
         Ok(client) => {
             let mut state = FELT_CLIENT.lock().expect("Could not lock mutex");
             trace!("firefox_connect_to_felt(): connected, storing client");
@@ -110,9 +122,9 @@ pub extern "C" fn felt_constructor(
     iid: *const xpcom::nsIID,
     result: *mut *mut xpcom::reexports::libc::c_void,
 ) -> nserror::nsresult {
-    let is_felt_ui = felt::IS_FELT_UI.load(Ordering::Relaxed);
-    let is_felt_browser = felt::IS_FELT_BROWSER.load(Ordering::Relaxed);
-    let felt_xpcom = felt::FeltXPCOM::new(is_felt_ui, is_felt_browser);
+    let is_felt_ui = crate::IS_FELT_UI.load(Ordering::Relaxed);
+    let is_felt_browser = crate::IS_FELT_BROWSER.load(Ordering::Relaxed);
+    let felt_xpcom = components::FeltXPCOM::new(is_felt_ui, is_felt_browser);
     unsafe { felt_xpcom.QueryInterface(iid, result) }
 }
 
@@ -121,6 +133,6 @@ pub extern "C" fn felt_restartforced_constructor(
     iid: *const xpcom::nsIID,
     result: *mut *mut xpcom::reexports::libc::c_void,
 ) -> nserror::nsresult {
-    let felt_restartforced = felt::FeltRestartForced::new();
+    let felt_restartforced = components::FeltRestartForced::new();
     unsafe { felt_restartforced.QueryInterface(iid, result) }
 }
