--- toolkit/library/felt/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "felt"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-cookie = "0.16"
-cstr = "0.2"
-ipc-channel = "0.19"
-log = "0.4"
-moz_task = { path = "../../../xpcom/rust/moz_task" }
-nsstring = { path = "../../../xpcom/rust/nsstring" }
-nserror = { path = "../../../xpcom/rust/nserror" }
-serde = { version = "1.0.219", features = ["derive"] }
-thin-vec = "0.2.12"
-time = "0.3.36"
-xpcom = { path = "../../../xpcom/rust/xpcom" }
-serde_json = "1.0"
