--- toolkit/library/felt/src/felt.h
@@ -1,12 +1,0 @@
-/* -*- Mode: C++; tab-width: 2; indent-tabs-mode: nil; c-basic-offset: 2 -*-
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#ifndef _felt_h_
-#define _felt_h_
-
-#include "../felt_ffi.h"
-
-#endif  // _felt_h_
