+++ toolkit/library/moz.build
@@ -139,9 +139,6 @@ DIRS += [
     "buildid_reader",
 ]
 
-if CONFIG["MOZ_ENTERPRISE"]:
-    DIRS += ["felt"]
-
 TEST_DIRS += ["gtest"]
 
 # js needs to come after xul for now, because it is an archive and its content
